Introduce provisioned state, don't check running supervisor unit.
There is a difference between being unprovisioned, provisioning and
provisioned. Clearly return once of these. Useful to avoid misreporting the
state to user while still being able to avoid duplicate provisioning attempts.

The detection of the supervisor unit being loaded/active is being removed for
now in this state check, it may be reintroduced as an additional field later.