Add workaround for prepare-openvpn.service issue.
prepare-openvpn.service reads config.json to obtain env vars, then uses these to
setup /etc/openvpn/resin.conf with the appropriate VPN endpoint. Unfortunately
it only reads this file on startup meaning if the endpoint is not available from
the beginning, the configuration will be incorrect.

We force a reload of prepare-openvpn.service (it's blocking so we don't wait for
the outcome) on supervisor enable/start at which point config.json should
contain the VPN endpoint correclty.